Split get-next-question.test.ts by call shape
